// shadcn/registry/default/data-table/types.ts
export interface DataTableRequestParams {
	page: number;
	pageSize: number;
	sorting: SortingState;
	columnFilters: ColumnFiltersState;
	globalFilter: string;
}